Initial implementation of querying server info

+++ ServerShared/Networking/ServerInfo.cs
@@ -5,7 +5,7 @@
         public string Name;
         public int Players;
         public int MaxPlayers;
-        public int Ping;
+        public float Ping;
 
         public string Ip;
         public int Port;
